Move codeeditor related code amd tests to own dir
