+++ spyder/plugins/editor/widgets/codeeditor/lsp_mixin.py
