+++ spyder/plugins/editor/widgets/codeeditor/tests/assets/autopep8_max_line.py
