+++ spyder/plugins/editor/widgets/codeeditor/tests/assets/autopep8_range.py
