+++ spyder/plugins/editor/widgets/codeeditor/tests/assets/black_max_line.py
