+++ spyder/plugins/editor/widgets/codeeditor/tests/assets/black_result.py
