+++ spyder/plugins/editor/widgets/codeeditor/tests/assets/braces.py
