+++ spyder/plugins/editor/widgets/codeeditor/tests/assets/original_file.py
