+++ spyder/plugins/editor/widgets/codeeditor/tests/assets/text.py
