+++ spyder/plugins/editor/widgets/codeeditor/tests/assets/yapf_max_line.py
