+++ spyder/plugins/editor/widgets/codeeditor/tests/assets/yapf_range.py
