+++ spyder/plugins/editor/widgets/codeeditor/tests/assets/yapf_result.py
