+++ spyder/plugins/editor/widgets/codeeditor/tests/test_autocolon.py
