+++ spyder/plugins/editor/widgets/codeeditor/tests/test_autoindent.py
