+++ spyder/plugins/editor/widgets/codeeditor/tests/test_breakpoints.py
