+++ spyder/plugins/editor/widgets/codeeditor/tests/test_classfunc_selector.py
