+++ spyder/plugins/editor/widgets/codeeditor/tests/test_comments.py
