+++ spyder/plugins/editor/widgets/codeeditor/tests/test_completions_hide.py
