+++ spyder/plugins/editor/widgets/codeeditor/tests/test_decorations.py
