+++ spyder/plugins/editor/widgets/codeeditor/tests/test_folding.py
