+++ spyder/plugins/editor/widgets/codeeditor/tests/test_goto.py
