+++ spyder/plugins/editor/widgets/codeeditor/tests/test_gotoline.py
