+++ spyder/plugins/editor/widgets/codeeditor/tests/test_hints_and_calltips.py
