+++ spyder/plugins/editor/widgets/codeeditor/tests/test_indentation.py
