+++ spyder/plugins/editor/widgets/codeeditor/tests/test_introspection.py
