+++ spyder/plugins/editor/widgets/codeeditor/tests/test_panels.py
