+++ spyder/plugins/editor/widgets/tests/test_editor_and_outline.py
@@ -10,14 +10,9 @@ Tests syncing between the EditorStack and OutlineExplorerWidget.
 
 # Standard library imports
 import os
-import json
 import os.path as osp
 import sys
 from unittest.mock import Mock
-
-# Qt imports
-from qtpy.QtCore import Qt
-from qtpy.QtGui import QTextCursor
 
 # Third party imports
 import pytest
@@ -25,21 +20,6 @@ import pytest
 # Local imports
 from spyder.config.base import running_in_ci
 from spyder.plugins.editor.widgets import editor
-from spyder.plugins.outlineexplorer.editor import OutlineExplorerProxyEditor
-from spyder.plugins.outlineexplorer.main_widget import OutlineExplorerWidget
-
-
-HERE = osp.dirname(osp.abspath(__file__))
-ASSETS = osp.join(HERE, 'assets')
-
-AVAILABLE_CASES = ['text']
-CASES = {
-    case: {
-        'file': osp.join(ASSETS, '{0}.py'.format(case)),
-        'tree': osp.join(ASSETS, '{0}_trees.json'.format(case))
-    }
-    for case in AVAILABLE_CASES
-}
 
 
 def get_tree_elements(treewidget):
@@ -83,37 +63,6 @@ def test_files(tmpdir_factory):
                 "# ---- a comment\n")
 
     return [filename1, filename2, filename3]
-
-
-@pytest.fixture
-def outlineexplorer(qtbot):
-    """Set up an OutlineExplorerWidget."""
-    outlineexplorer = OutlineExplorerWidget(None, None, None)
-    outlineexplorer.set_conf('show_fullpath', False)
-    outlineexplorer.set_conf('show_all_files', True)
-    outlineexplorer.set_conf('group_cells', True)
-    outlineexplorer.set_conf('show_comments', True)
-    outlineexplorer.set_conf('sort_files_alphabetically', False)
-    outlineexplorer.set_conf('display_variables', True)
-
-    # Fix the size of the outline explorer to prevent an
-    # 'Unable to set geometry ' warning if the test fails.
-    outlineexplorer.setFixedSize(400, 350)
-
-    qtbot.addWidget(outlineexplorer)
-    outlineexplorer.show()
-
-    return outlineexplorer
-
-
-@pytest.fixture
-def completions_codeeditor_outline(completions_codeeditor, outlineexplorer):
-    editor, _ = completions_codeeditor
-    editor.oe_proxy = OutlineExplorerProxyEditor(editor, editor.filename)
-    outlineexplorer.register_editor(editor.oe_proxy)
-    outlineexplorer.set_current_editor(
-        editor.oe_proxy, update=False, clear=False)
-    return editor, outlineexplorer
 
 
 @pytest.fixture
@@ -286,170 +235,5 @@ def test_toggle_on_show_all_files(editorstack, outlineexplorer, test_files):
     assert results == ['text1.txt', 'foo1.py', 'foo2.py']
 
 
-@pytest.mark.order(2)
-def test_editor_outlineexplorer(qtbot, completions_codeeditor_outline):
-    """Tests that the outline explorer reacts to editor changes."""
-    code_editor, outlineexplorer = completions_codeeditor_outline
-    treewidget = outlineexplorer.treewidget
-    treewidget.is_visible = True
-
-    case_info = CASES['text']
-    filename = case_info['file']
-    tree_file = case_info['tree']
-
-    with open(filename, 'r') as f:
-        lines = f.read()
-
-    with open(tree_file, 'r') as f:
-        trees = json.load(f)
-
-    code_editor.toggle_automatic_completions(False)
-    code_editor.toggle_code_snippets(False)
-
-    # Set cursor to start
-    code_editor.set_text('')
-    code_editor.go_to_line(1)
-
-    # Put example text in editor
-    code_editor.set_text(lines)
-    qtbot.wait(3000)
-
-    # Check that the outline tree was initialized successfully
-    tree = trees[0]
-    root_tree = get_tree_elements(treewidget)
-    assert root_tree == tree
-
-    # Remove "d" symbol
-    code_editor.go_to_line(14)
-    cursor = code_editor.textCursor()
-    start = code_editor.get_position_line_number(13, -1)
-    end = code_editor.get_position_line_number(17, 0)
-    cursor.setPosition(start)
-    cursor.setPosition(end, QTextCursor.KeepAnchor)
-    code_editor.setTextCursor(cursor)
-    code_editor.cut()
-    qtbot.wait(3000)
-
-    tree = trees[1]
-    root_tree = get_tree_elements(treewidget)
-    assert root_tree == tree
-
-    # Add "d" symbol elsewhere
-    code_editor.go_to_line(36)
-
-    with qtbot.waitSignal(
-            code_editor.completions_response_signal, timeout=30000):
-        qtbot.keyPress(code_editor, Qt.Key_Return)
-        qtbot.keyPress(code_editor, Qt.Key_Return)
-
-    qtbot.keyPress(code_editor, Qt.Key_Up)
-    code_editor.paste()
-
-    qtbot.wait(3000)
-
-    tree = trees[2]
-    root_tree = get_tree_elements(treewidget)
-    assert root_tree == tree
-
-    # Move method1
-    code_editor.go_to_line(56)
-    cursor = code_editor.textCursor()
-    start = code_editor.get_position_line_number(55, -1)
-    end = code_editor.get_position_line_number(57, -1)
-    cursor.setPosition(start)
-    cursor.setPosition(end, QTextCursor.KeepAnchor)
-    code_editor.setTextCursor(cursor)
-    code_editor.cut()
-    qtbot.wait(3000)
-
-    tree = trees[3]
-    root_tree = get_tree_elements(treewidget)
-    assert root_tree == tree
-
-    # Add method1
-    code_editor.go_to_line(49)
-
-    with qtbot.waitSignal(
-            code_editor.completions_response_signal, timeout=30000):
-        qtbot.keyPress(code_editor, Qt.Key_Return)
-        qtbot.keyPress(code_editor, Qt.Key_Return)
-
-    qtbot.keyPress(code_editor, Qt.Key_Up)
-    code_editor.paste()
-    qtbot.wait(3000)
-
-    tree = trees[4]
-    root_tree = get_tree_elements(treewidget)
-    assert root_tree == tree
-
-    # Add attribute "y"
-    code_editor.go_to_line(48)
-    cursor = code_editor.textCursor()
-    cursor.movePosition(QTextCursor.EndOfBlock)
-    code_editor.setTextCursor(cursor)
-
-    with qtbot.waitSignal(
-            code_editor.completions_response_signal, timeout=30000):
-        qtbot.keyPress(code_editor, Qt.Key_Return)
-        qtbot.keyClicks(code_editor, 'self.y = None')
-        qtbot.keyPress(code_editor, Qt.Key_Return)
-
-    with qtbot.waitSignal(treewidget.sig_tree_updated, timeout=30000):
-        code_editor.request_symbols()
-
-    tree = trees[5]
-    root_tree = get_tree_elements(treewidget)
-    assert root_tree == tree
-
-
-@pytest.mark.order(2)
-def test_empty_file(qtbot, completions_codeeditor_outline):
-    """
-    Test that the outline explorer is updated correctly when
-    it's associated file is empty.
-    """
-    code_editor, outlineexplorer = completions_codeeditor_outline
-    treewidget = outlineexplorer.treewidget
-    treewidget.is_visible = True
-
-    code_editor.toggle_automatic_completions(False)
-    code_editor.toggle_code_snippets(False)
-
-    # Set empty contents
-    code_editor.set_text('')
-    code_editor.go_to_line(1)
-    qtbot.wait(3000)
-
-    # Assert the spinner is not shown.
-    assert not outlineexplorer._spinner.isSpinning()
-
-    # Add some content
-    code_editor.set_text("""
-def foo():
-    a = 10
-    return a
-""")
-    qtbot.wait(3000)
-
-    root_tree = get_tree_elements(treewidget)
-    assert root_tree == {'test.py': [{'foo': [{'a': []}]}]}
-
-    # Remove content
-    code_editor.selectAll()
-    qtbot.keyPress(code_editor, Qt.Key_Delete)
-
-    with qtbot.waitSignal(
-            code_editor.completions_response_signal, timeout=30000):
-        code_editor.document_did_change()
-
-    qtbot.wait(3000)
-
-    # Assert the tree is empty and the spinner is not shown.
-    root_tree = get_tree_elements(treewidget)
-    assert root_tree == {'test.py': []}
-    assert not outlineexplorer._spinner.isSpinning()
-
-
 if __name__ == "__main__":
-    import os
     pytest.main([os.path.basename(__file__), '-vv', '-rw'])
